Replacing user branch: prevent by default (#27)

+++ Modules/AbRouter/Services/Experiment/UserExperimentsService.php
@@ -95,7 +95,7 @@ class UserExperimentsService
                 $experimentUser->id
             );
 
-        if ($experimentBranchUser) {
+        if ($experimentBranchUser && $userExperimentDTO->isForce()) {
             $experimentBranchUser->delete();
         }
